+++ .github/workflows/test.yml
@@ -9,6 +9,27 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
+  unit-tests:
+    name: Unit Tests
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+
+      - name: Install dependencies
+        working-directory: cli
+        run: bun install
+
+      - name: Run bun tests
+        run: bun test
+
+      - name: Verify cloud bundles build
+        run: bun run cli/build-clouds.ts
+
   mock-tests:
     name: Mock Tests
     runs-on: ubuntu-latest
@@ -16,6 +37,13 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+
+      - name: Install dependencies
+        working-directory: cli
+        run: bun install
 
       - name: Run mock tests
         id: tests
