+++ .gitignore
@@ -3,6 +3,7 @@ node_modules/
 __pycache__/
 .claude/skills/*/start-*.sh
 cli/cli.js
+cli/fly.js
 
 # Sensitive files — never commit secrets or private keys
 .env
