+++ cli/src/__tests__/cloud-lib-source-chain.test.ts
@@ -123,8 +123,9 @@ describe("shared/common.sh prerequisite", () => {
 // ── Cloud lib source chain ──────────────────────────────────────────────────
 
 describe("Cloud lib/common.sh source chain", () => {
-  it(`should discover at least 9 cloud lib files`, () => {
-    expect(allClouds.length).toBeGreaterThanOrEqual(8);
+  it(`should discover at least 7 cloud lib files`, () => {
+    // Note: TS-based clouds (e.g. fly) don't have bash lib/common.sh
+    expect(allClouds.length).toBeGreaterThanOrEqual(7);
   });
 
   for (const cloud of allClouds) {
